Use long as offset type within Struct class

+++ src/main/java/de/hhu/bsinfo/neutrino/data/NativeByte.java
@@ -5,7 +5,7 @@ import java.nio.ByteBuffer;
 
 public class NativeByte extends NativeDataType {
 
-    public NativeByte(final LocalBuffer byteBuffer, final int offset) {
+    public NativeByte(final LocalBuffer byteBuffer, final long offset) {
         super(byteBuffer, offset);
     }
