+++ src/main/java/de/hhu/bsinfo/neutrino/data/NativeString.java
@@ -10,7 +10,7 @@ public class NativeString extends NativeDataType {
 
     private static final byte ZERO = 0;
 
-    public NativeString(final LocalBuffer byteBuffer, final int offset, final int length) {
+    public NativeString(final LocalBuffer byteBuffer, final long offset, final int length) {
         super(byteBuffer, offset);
         this.length = length;
     }
