+++ src/main/java/de/hhu/bsinfo/neutrino/util/ValueFactory.java
@@ -6,5 +6,5 @@ import java.nio.ByteBuffer;
 
 @FunctionalInterface
 public interface ValueFactory<T extends NativeObject> {
-    T newInstance(LocalBuffer byteBuffer, int offset);
+    T newInstance(LocalBuffer byteBuffer, long offset);
 }
